requirements.txt
fastapi
pydantic
dotenv
mysql-connector-python